Ensure systemd-sysv package is installed in debian-based images
The systemd-sysv package does not get installed along with systemd on
some versions of Debian (9, 11) and Ubuntu (20.04). This simply brings
those images in line with the others, and ensures `/sbin/init` is a
symlink to systemd.

Looking at `/sbin/init` and where it symlinks to is one way to determine
what init-system is used. Ansible's service_mgr module for example does
just that:

https://github.com/ansible/ansible/blob/3acd8f6f7f61b72a42a017060c3a718bbbeaf854/lib/ansible/module_utils/facts/system/service_mgr.py#L61